+++ frontend/app/experiments/(experiment)/JobDetail.tsx
