+++ frontend/app/experiments/(experiment)/[flowID]/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function ExperimentDetail() {
+  return null;
+}
