+++ frontend/app/experiments/(experiment)/new/[taskSlug]/[[...toolCID]]/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function NewExperiment() {
+  return null;
+}
