+++ frontend/components/visualization/Molstar/skin.scss
@@ -1,4 +1,4 @@
-$default-background: #ffffff;
+$default-background: #f9fafb;
 $font-color: #000000;
 $hover-font-color: #000000;
 $entity-current-font-color: #0343ba;
